build: switch to meson
The original codebase didn't support compiling on NixOS very well, due
to differences in finding the headers for the MOTIF and X11 Toolkit
Intrinsic libraries, as well as the linker flags needed to link to them.
While a bit overkill for compiling a single C++ file, this at least
provides users the standardized options that the original makefile did
not provide.